Fix Tooltip position when root view is not located at (0, 0)
Element position is measured relatively to root view which causes
Tooltip to render at wrong place inside the Modal when root view is not
starting at (0, 0) coordinates of window. Measuring element position in
window fixes this issue as Modal always start at (0, 0) and is
fullscreen.

Rest of the changes are forces by prettier and linter.